# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sistema de Gerenciamento Multi-Empresarial de Documentos Fiscais",
        "title": "ZoomXML API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.zoomxml.com/support",
            "email": "support@zoomxml.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "AdminToken": []
                    }
                ],
                "description": "Lista todos os usuários do sistema com paginação e filtros",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Listar usuários",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrar por papel (admin/user)",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por status (true/false)",
                        "name": "active",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Página (padrão: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Itens por página (padrão: 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários com paginação",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerUsersResponse"
                        }
                    },
                    "401": {
                        "description": "Token de admin necessário",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminToken": []
                    }
                ],
                "description": "Cria um novo usuário no sistema (apenas admin com token especial)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Criar usuário",
                "parameters": [
                    {
                        "description": "Dados do usuário",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerUserWithToken"
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerValidationError"
                        }
                    },
                    "401": {
                        "description": "Token de admin necessário",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerError"
                        }
                    },
                    "409": {
                        "description": "Email já existe",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "password": {
                    "description": "Senha obrigatória para frontend",
                    "type": "string",
                    "minLength": 8
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "user"
                    ]
                },
                "token": {
                    "description": "Token opcional - se não fornecido, será gerado automaticamente",
                    "type": "string"
                }
            }
        },
...
        "handlers.SwaggerError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Mensagem de erro"
                }
            }
        },
        "handlers.SwaggerPagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "handlers.SwaggerUser": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-08-17T19:01:44Z"
                },
                "email": {
                    "type": "string",
                    "example": "joao@exemplo.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "João Silva"
                },
                "role": {
                    "type": "string",
                    "example": "user"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-08-17T19:01:44Z"
                }
            }
        },
        "handlers.SwaggerUserWithToken": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-08-17T19:01:44Z"
                },
                "email": {
                    "type": "string",
                    "example": "joao@exemplo.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "João Silva"
                },
                "role": {
                    "type": "string",
                    "example": "user"
                },
                "token": {
                    "type": "string",
                    "example": "U6HGHy4SDK"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-08-17T19:01:44Z"
                }
            }
        },
        "handlers.SwaggerUsersResponse": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/handlers.SwaggerPagination"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.SwaggerUser"
                    }
                }
            }
        },
        "handlers.SwaggerValidationError": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "{\"email\"": "\"email is required\"}"
                    }
                },
                "error": {
                    "type": "string",
                    "example": "Validation failed"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminToken": {
            "description": "Token de administrador para operações de usuários (ex: admin-secret-token)",
            "type": "apiKey",
            "name": "token",
            "in": "header"
        },
        "UserToken": {
            "description": "Token de usuário para autenticação (ex: U6HGHy4SDK)",
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}